Implement `BaseWandbTrainingCallback`: - Turn `WandbCustomCallback` into an abstract class - Create `WandbFinetuneCallback` and `WandbDistillationCallback`
